requirements.txt
aiohttp==1.0.5
async-timeout==3.0.1
cffi==1.11.5
chardet==3.0.4
discord==0.0.2
discord.py==0.16.12
multidict==4.5.2
pycparser==2.19
PyNaCl==1.3.0
python-decouple==3.1
six==1.12.0
websockets==3.4
